Convert cost calculation to long instead of int after rounding in status bar.
This was an issue for some dropships that were showing a negative cost
in the status bar, but I went ahead and made the change for all units.

+++ src/megameklab/com/ui/Mek/StatusBar.java
@@ -122,7 +122,7 @@ public class StatusBar extends ITab {
             maxCrits = 78;
         }
         int currentCrits = UnitUtil.countUsedCriticals(getMech());
-        int currentCost = (int)Math.round(getMech().getCost(false));
+        long currentCost = (long) Math.round(getMech().getCost(false));
 
         testEntity = new TestMech(getMech(), entityVerifier.mechOption, null);
